Improve playout

+++ source/tesuji/games/go/monte_carlo/AbstractMonteCarloAdministration.java
@@ -769,9 +769,9 @@ public abstract class AbstractMonteCarloAdministration
      * (non-Javadoc)
      * @see tesuji.games.go.monte_carlo.MonteCarloAdministration#getScore()
      */
-    public double getScore()
+    public int getScore()
     {
-    	double score = getScoreEstimate();
+    	int score = getScoreEstimate();
     	
     	for (int i=_emptyPoints.getSize(); --i>=0;)
     	{
@@ -797,9 +797,9 @@ public abstract class AbstractMonteCarloAdministration
     /**
      * @return a very rough estimate of the score, counting just the stones and komi.
      */
-    private double getScoreEstimate()
+    private int getScoreEstimate()
     {
-    	return (double)_nrBlackStones - (double)_nrWhiteStones - _komi;
+    	return _nrBlackStones - _nrWhiteStones - (int)_komi;
     }
     
     /*
